+++ src/components/ui/input.tsx
@@ -22,13 +22,13 @@ const inputVariants = cva(
   {
     variants: {
       size: {
-        default: "h-9 px-3 py-1 text-sm file:text-sm",
+        md: "h-9 px-3 py-1 text-sm file:text-sm",
         sm: "h-8 px-2.5 py-1 text-xs file:text-xs",
         xs: "h-7 px-2 py-0.5 text-xs file:text-xs",
       },
     },
     defaultVariants: {
-      size: "default",
+      size: "md",
     },
   },
 );